minor stuff before bed...
